Firmware which allows you to write only to SD card
Introduce the 'N' command, which causes only writing to
SD card. This is useful to avoid the general problem of the
baudrate slowdown from the RFDuino, which causes you not
to be able to run the ADS1299 at a higher sample rate.

Note that this does *not* allow you to do things in real time
as that would require board mods. But it does at least allow you
to collect data.